Revert "Snapshot: Fixing Task Journey Reporting and Error Handling"
This reverts commit 689dab2bd4843c41167b645b178caae5a2604642.
