+++ controller/src/main/java/net/fhirfactory/pegacorn/ponos/subsystem/processingplant/PonosAcolyte.java
@@ -21,8 +21,7 @@
  */
 package net.fhirfactory.pegacorn.ponos.subsystem.processingplant;
 
-import net.fhirfactory.pegacorn.core.model.topology.valuesets.ProcessingPlantProviderRoleEnum;
-import net.fhirfactory.pegacorn.core.model.topology.valuesets.ProcessingPlantTypeEnum;
+import net.fhirfactory.pegacorn.core.model.topology.role.ProcessingPlantRoleEnum;
 import net.fhirfactory.pegacorn.ponos.interfaces.PonosSubsystemDetailsInterface;
 import net.fhirfactory.pegacorn.ponos.workshops.workflow.monitoring.TaskFulfillmentWatchdog;
 import net.fhirfactory.pegacorn.processingplant.ProcessingPlant;
@@ -51,18 +50,14 @@ public abstract class PonosAcolyte extends ProcessingPlant implements PonosSubsy
     //
 
     @Override
-    protected ProcessingPlantProviderRoleEnum specifyPlantProviderRole() {
-        return ProcessingPlantProviderRoleEnum.PETASOS_SERVICE_PROVIDER_TASK_MANAGEMENT;
+    public ProcessingPlantRoleEnum getProcessingPlantCapability() {
+        return (ProcessingPlantRoleEnum.PETASOS_SERVICE_PROVIDER_TASK_MANAGEMENT);
     }
+
     @Override
     public String getInfinispanJGroupsConfigFile() {
         String multiuseInfinispanStackConfigFile = specifyPropertyFile().getDeploymentMode().getMultiuseInfinispanStackConfigFile();
         return (multiuseInfinispanStackConfigFile);
-    }
-
-    @Override
-    protected ProcessingPlantTypeEnum specifyProcessingPlantType() {
-        return (ProcessingPlantTypeEnum.PROCESSING_PLANT_TYPE_INFORMATION_MANAGER);
     }
 
     //
