+++ controller/src/main/java/net/fhirfactory/pegacorn/ponos/workshops/workflow/status/activities/AggregateTaskReportingActivities.java
@@ -22,10 +22,8 @@
 package net.fhirfactory.pegacorn.ponos.workshops.workflow.status.activities;
 
 import net.fhirfactory.pegacorn.core.constants.petasos.PetasosPropertyConstants;
-import net.fhirfactory.pegacorn.core.model.component.valuesets.SoftwareComponentConnectivityContextEnum;
 import net.fhirfactory.pegacorn.core.model.componentid.ComponentIdType;
-import net.fhirfactory.pegacorn.core.model.petasos.oam.notifications.ITOpsNotification;
-import net.fhirfactory.pegacorn.core.model.petasos.oam.notifications.ITOpsNotificationContent;
+import net.fhirfactory.pegacorn.core.model.petasos.oam.notifications.PetasosComponentITOpsNotification;
 import net.fhirfactory.pegacorn.core.model.petasos.task.PetasosActionableTask;
 import net.fhirfactory.pegacorn.core.model.petasos.task.datatypes.fulfillment.valuesets.FulfillmentExecutionStatusEnum;
 import net.fhirfactory.pegacorn.core.model.petasos.task.datatypes.identity.datatypes.TaskIdType;
@@ -36,9 +34,8 @@ import net.fhirfactory.pegacorn.internals.fhir.r4.internal.topics.HL7V2XTopicFac
 import net.fhirfactory.pegacorn.petasos.core.tasks.factories.metadata.GeneralTaskMetadataExtractor;
 import net.fhirfactory.pegacorn.petasos.core.tasks.factories.metadata.HL7v2xTaskMetadataExtractor;
 import net.fhirfactory.pegacorn.petasos.oam.metrics.agents.ProcessingPlantMetricsAgentAccessor;
-import net.fhirfactory.pegacorn.ponos.reporitng.factories.AggregateTaskReportContentExtractor;
 import net.fhirfactory.pegacorn.ponos.workshops.oam.ProcessingPlantTaskReportProxy;
-import net.fhirfactory.pegacorn.ponos.reporitng.factories.AggregateTaskReportFactory;
+import net.fhirfactory.pegacorn.ponos.workshops.workflow.factories.AggregateTaskReportFactory;
 import net.fhirfactory.pegacorn.ponos.workshops.workflow.status.activities.common.TaskActivityProcessorBase;
 import net.fhirfactory.pegacorn.ponos.workshops.datagrid.cache.PonosPetasosActionableTaskCacheServices;
 import org.apache.commons.lang3.exception.ExceptionUtils;
@@ -50,10 +47,11 @@ import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Timer;
 import java.util.TimerTask;
-
+import net.fhirfactory.pegacorn.ponos.workshops.workflow.factories.EndpointInformationExtractor;
 import org.apache.commons.lang3.StringUtils;
 
 @ApplicationScoped
@@ -94,7 +92,7 @@ public class AggregateTaskReportingActivities extends TaskActivityProcessorBase 
     private SubsystemNames subsystemNames;
 
     @Inject
-    private AggregateTaskReportContentExtractor endpointInfoExtrator;
+    private EndpointInformationExtractor endpointInfoExtrator;
 
     //
     // Constructor(s)
@@ -172,8 +170,8 @@ public class AggregateTaskReportingActivities extends TaskActivityProcessorBase 
                         try {
                             getLogger().debug(".aggregateTaskReportingDaemon(): Iterating, currentTask->{}", currentTask.getTaskId());
                             if (!currentTask.getTaskFulfillment().getFulfillerWorkUnitProcessor().getSubsystemParticipantName().contains(subsystemNames.getITOpsIMParticipantName())) {
-                                publishEndOfChainDetailedTaskReport(currentTask);
-                                publishEndOfChainSummaryTaskReport(currentTask);
+                                publishEndOfChainFullTaskReport(currentTask);
+                                publishEndPointOnlyTaskReport(currentTask);
                             }
                             getTaskCacheServices().archivePetasosActionableTask(currentTask);
                             getTaskCacheServices().setReportStatus(currentTask.getTaskId(), true);
@@ -193,31 +191,10 @@ public class AggregateTaskReportingActivities extends TaskActivityProcessorBase 
         getLogger().debug(".aggregateTaskReportingDaemon(): Exit");
     }
 
-    public void publishEndOfChainDetailedTaskReport(PetasosActionableTask lastTask) {
-        getLogger().debug(".publishEndOfChainDetailedTaskReport(): Entry, lastTask->{}", lastTask);
+    public void publishEndOfChainFullTaskReport(PetasosActionableTask lastTask) {
+        getLogger().debug(".publishEndOfChainTaskReport(): Entry, lastTask->{}", lastTask);
 
         TaskTraceabilityType taskTraceability = lastTask.getTaskTraceability();
-
-        boolean isEdge = false;
-        try {
-            isEdge = lastTask.getTaskFulfillment().getFulfillerWorkUnitProcessor().equals(SoftwareComponentConnectivityContextEnum.COMPONENT_ROLE_SUBSYSTEM_EDGE);
-        } catch(Exception Ex){
-            getLogger().warn(".publishEndOfChainSummaryTaskReport(): Fulfiller WorkUnitProcessor does not have a role defined!");
-        }
-        //
-        // Check to see if it isn't an internal (+ve) report for message forwarding
-        if (isEdge) {
-            // we don't want to send these if they are OK outcomes
-            try {
-                if (lastTask.getTaskFulfillment().getStatus().equals(FulfillmentExecutionStatusEnum.FULFILLMENT_EXECUTION_STATUS_FINISHED) || lastTask.getTaskFulfillment().getStatus().equals(FulfillmentExecutionStatusEnum.FULFILLMENT_EXECUTION_STATUS_FINALISED)) {
-                    // exit out
-                    return;
-                }
-            } catch (Exception ex) {
-                getLogger().warn(".publishEndOfChainSummaryTaskReport(): Couldn't resolve task outcome status, message->{}", ExceptionUtils.getMessage(ex));
-                return;
-            }
-        }
 
         //
         // Get the 1st
@@ -227,43 +204,52 @@ public class AggregateTaskReportingActivities extends TaskActivityProcessorBase 
 
         //
         // Create the Report
-        ITOpsNotification notification = getAggregateTaskReportFactory().endOfChainReport(lastTask);
-        String reportString = notification.getContent();
-        String formattedReportString = notification.getFormattedContent();
+        String reportString = getAggregateTaskReportFactory().endOfChainReport(lastTask);
 
         //
-        // Resolve Endpoint Component ID(s)
-        ComponentIdType egressComponentId = getEndpointInfoExtrator().getEndpointComponentId(lastTask);
-        ComponentIdType ingresComponentId = getEndpointInfoExtrator().getEndpointComponentId(firstTask);
+        // Publish to Last Participant
+        String lastSubsystemName = lastTask.getTaskFulfillment().getFulfillerWorkUnitProcessor().getSubsystemParticipantName();
+        ComponentIdType lastComponentId = lastTask.getTaskFulfillment().getFulfillerWorkUnitProcessor().getComponentID();
+        taskReportProxy.sendITOpsTaskReport(lastSubsystemName,lastComponentId,reportString);
 
         //
-        // Resolve Ingres Endpoint Participant Name(s)
-        String ingresEndpointParticipantName = getEndpointInfoExtrator().getEndpointParticipantName(firstTask, true);
+        // Publish to First Participant
+        String firstSubsystemName = firstTask.getTaskFulfillment().getFulfillerWorkUnitProcessor().getSubsystemParticipantName();
+        ComponentIdType firstComponentId = firstTask.getTaskFulfillment().getFulfillerWorkUnitProcessor().getComponentID();
+        taskReportProxy.sendITOpsTaskReport(firstSubsystemName,firstComponentId, reportString);
+
+        //
+        // All done
+        getLogger().debug(".publishEndOfChainTaskReport(): Exit");
+    }
+
+    public void publishEndPointOnlyTaskReport(PetasosActionableTask lastTask){
+        getLogger().debug(".publishEndPointOnlyTaskReport(): Entry, lastTask->{}", lastTask);
+
+        //
+        // Create the report
+        PetasosComponentITOpsNotification notificationBase = getAggregateTaskReportFactory().newEndpointOnlyTaskReport(lastTask);
+
         //
         // Resolve Egress Endpoint Participant Name(s)
         String egressEndpointParticipantName = getEndpointInfoExtrator().getEndpointParticipantName(lastTask, false);
 
-
-        if(StringUtils.isNotEmpty(egressEndpointParticipantName) && egressComponentId != null ){
-            taskReportProxy.sendITOpsEndpointOnlyTaskReport(egressEndpointParticipantName,egressComponentId, reportString, formattedReportString);
-        }
-
-        if(StringUtils.isNotEmpty(ingresEndpointParticipantName) && ingresComponentId != null ){
-            taskReportProxy.sendITOpsEndpointOnlyTaskReport(ingresEndpointParticipantName,ingresComponentId, reportString, formattedReportString);
-        }
-
         //
-        // All done
-        getLogger().debug(".publishEndOfChainDetailedTaskReport(): Exit");
-    }
-
-
-    public void publishEndOfChainSummaryTaskReport(PetasosActionableTask lastTask){
-        getLogger().debug(".publishEndOfChainSummaryTaskReport(): Entry, lastTask->{}", lastTask);
-
-        //
-        // Create the report
-        ITOpsNotificationContent notificationContent = getAggregateTaskReportFactory().newEndpointOnlyTaskReport(lastTask);
+        // Check to see if it isn't an internal (+ve) report for message forwarding
+        if(StringUtils.isNotEmpty(egressEndpointParticipantName)) {
+            if (egressEndpointParticipantName.contains("PetasosTaskForwarderWUP")) {
+                // we don't want to send these if they are OK outcomes
+                try {
+                    if (lastTask.getTaskFulfillment().getStatus().equals(FulfillmentExecutionStatusEnum.FULFILLMENT_EXECUTION_STATUS_FINISHED) || lastTask.getTaskFulfillment().getStatus().equals(FulfillmentExecutionStatusEnum.FULFILLMENT_EXECUTION_STATUS_FINALISED)) {
+                        // exit out
+                        return;
+                    }
+                } catch (Exception ex) {
+                    getLogger().warn(".publishEndPointOnlyTaskReport(): Couldn't resolve task outcome status, message->{}", ExceptionUtils.getMessage(ex));
+                    return;
+                }
+            }
+        }
 
         //
         // Get the 1st Task
@@ -273,18 +259,23 @@ public class AggregateTaskReportingActivities extends TaskActivityProcessorBase 
         PetasosActionableTask firstTask = getTaskCacheServices().getPetasosActionableTask(firstTaskId);
 
         //
-        // Publish to Last Participant
-        String lastSubsystemName = lastTask.getTaskFulfillment().getFulfillerWorkUnitProcessor().getSubsystemParticipantName();
-        ComponentIdType lastComponentId = lastTask.getTaskFulfillment().getFulfillerWorkUnitProcessor().getComponentID();
-        taskReportProxy.sendITOpsTaskReport(lastSubsystemName,lastComponentId,notificationContent.getContent(), notificationContent.getFormattedContent());
+        // Resolve Ingres Endpoint Participant Name(s)
+        String ingresEndpointParticipantName = getEndpointInfoExtrator().getEndpointParticipantName(firstTask, true);
 
         //
-        // Publish to First Participant
-        String firstSubsystemName = firstTask.getTaskFulfillment().getFulfillerWorkUnitProcessor().getSubsystemParticipantName();
-        ComponentIdType firstComponentId = firstTask.getTaskFulfillment().getFulfillerWorkUnitProcessor().getComponentID();
-        taskReportProxy.sendITOpsTaskReport(firstSubsystemName,firstComponentId, notificationContent.getContent(), notificationContent.getFormattedContent());
+        // Resolve Endpoint Component ID(s)
+        ComponentIdType egressComponentId = getEndpointInfoExtrator().getEndpointComponentId(lastTask);
+        ComponentIdType ingresComponentId = getEndpointInfoExtrator().getEndpointComponentId(firstTask);
 
-        getLogger().debug(".publishEndOfChainSummaryTaskReport(): Exit, report->{}", notificationContent);
+        if(StringUtils.isNotEmpty(egressEndpointParticipantName) && egressComponentId != null ){
+            taskReportProxy.sendITOpsEndpointOnlyTaskReport(egressEndpointParticipantName,egressComponentId, notificationBase.getContent(), notificationBase.getFormattedContent());
+        }
+
+        if(StringUtils.isNotEmpty(ingresEndpointParticipantName) && ingresComponentId != null ){
+            taskReportProxy.sendITOpsEndpointOnlyTaskReport(ingresEndpointParticipantName,ingresComponentId, notificationBase.getContent(), notificationBase.getFormattedContent());
+        }
+
+        getLogger().debug(".publishEndPointOnlyTaskReport(): Exit, report->{}", notificationBase);
     }
 
     //
@@ -319,7 +310,7 @@ public class AggregateTaskReportingActivities extends TaskActivityProcessorBase 
         return(aggregateTaskReportFactory);
     }
 
-    protected AggregateTaskReportContentExtractor getEndpointInfoExtrator(){
+    protected EndpointInformationExtractor getEndpointInfoExtrator(){
         return(this.endpointInfoExtrator);
     }
