+++ dgis/src/main/java/net/fhirfactory/pegacorn/ponos/workshops/datagrid/cache/PonosPetasosParticipantCacheServices.java
@@ -202,14 +202,14 @@ public class PonosPetasosParticipantCacheServices {
      * @return
      */
     public PetasosParticipantRegistration updatePetasosParticipant(PetasosParticipant participant){
-        getLogger().info(".updatePetasosParticipant(): Entry, participant->{}", participant);
+        getLogger().debug(".updatePetasosParticipant(): Entry, participant->{}", participant);
         if(participant == null){
             getLogger().debug(".updatePetasosParticipant(): Exit, participant is null, returning null");
             return(null);
         }
         deregisterPetasosParticipant(participant.getComponentID());
         PetasosParticipantRegistration registration = registerPetasosParticipant(participant);
-        getLogger().info(".updatePetasosParticipant(): Exit, registration->{}", registration);
+        getLogger().debug(".updatePetasosParticipant(): Exit, registration->{}", registration);
         return(registration);
     }
